Adds basic health system and hit detection
Changed the P1Loc and P2Loc to be more than placeholders and act as
the object to refer to for all needs referring to the according player.
PlayerScript will maintain all the information about the player such as
health and other attributes such as paint resources remaining(to be
added later). Use/add getter/setter methods to PlayerScript to changes
attributes to make changes uniform and clear (will help with
networking). From now on, put any objects that refers to one of the
players under the player prefab/object.

Misc. Made a scene for testing collision; Nevermind it, for personal
use for testing.